# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/boards/{boardId}/cards/{cardId}/dependencies": {
            "get": {
                "description": "Retrieves the dependencies of a specific kanban card within a board.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Get card dependencies",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Card ID",
                        "name": "cardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of card dependencies",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/card.Card"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid board or card ID",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    },
                    "404": {
                        "description": "Board or card not found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates dependencies for a kanban card by specifying dependent card IDs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Create card dependencies",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Card ID",
                        "name": "cardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Array of dependency card IDs",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid input or missing dependency card IDs",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    },
                    "404": {
                        "description": "Board or card not found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes all dependencies for the specified card in the given board.",
                "tags": [
                    "cards"
                ],
                "summary": "Delete all dependencies of a kanban card",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Card ID",
                        "name": "cardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid board or card ID",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    },
                    "404": {
                        "description": "Board or card not found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    }
                }
            }
        },
        "/boards/{boardId}/cards/{cardId}/dependencies/{cardDependencyId}": {
            "delete": {
                "description": "Deletes a dependency from a kanban card by its ID.",
                "tags": [
                    "cards"
                ],
                "summary": "Delete a card dependency",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Card ID",
                        "name": "cardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Card Dependency ID",
                        "name": "cardDependencyId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorMsg"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "card.Card": {
            "type": "object",
            "required": [
                "acceptanceCriteria",
                "description",
                "status",
                "storyPoints",
                "title",
                "type"
            ],
            "properties": {
                "acceptanceCriteria": {
                    "type": "string",
                    "minLength": 3
                },
                "assignee": {
                    "type": "string"
                },
                "blockedReason": {
                    "type": "string"
                },
                "boardId": {
                    "description": "should this be at this level just id or the whole board document",
                    "type": "integer"
                },
                "cardId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "minLength": 3
                },
                "isBlocked": {
                    "type": "boolean"
                },
                "priority": {
                    "$ref": "#/definitions/card.Priority"
                },
                "status": {
                    "minLength": 3,
                    "allOf": [
                        {
                            "$ref": "#/definitions/card.Status"
                        }
                    ]
                },
                "storyPoints": {
                    "type": "integer"
                },
                "timeSpentHours": {
                    "type": "number"
                },
                "title": {
                    "type": "string",
                    "minLength": 3
                },
                "type": {
                    "minLength": 3,
                    "allOf": [
                        {
                            "$ref": "#/definitions/card.Type"
                        }
                    ]
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                }
            }
        },
        "card.Priority": {
            "type": "string",
            "enum": [
                "High",
                "Medium",
                "Low"
            ],
            "x-enum-varnames": [
                "PriorityHigh",
                "PriorityMedium",
                "PriorityLow"
            ]
        },
        "card.Status": {
            "type": "string",
            "enum": [
                "To Do",
                "In Progress",
                "In Review",
                "Done"
            ],
            "x-enum-varnames": [
                "StatusTodo",
                "StatusInProgress",
                "StatusInReview",
                "StatusDone"
            ]
        },
        "card.Type": {
            "type": "string",
            "enum": [
                "Bug",
                "Feature",
                "Maintenance"
            ],
            "x-enum-varnames": [
                "TypeBug",
                "TypeFeature",
                "TypeMaintenance"
            ]
        },
...
        "common.ErrorMsg": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}